Change plash screen behavior
