+++ public/index.html
@@ -28,6 +28,6 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
+    <div id="root" style="overflow: hidden;"></div>
   </body>
 </html>
