+++ src/index.js
@@ -3,10 +3,13 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { HashRouter } from 'react-router-dom';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render( 
-    <App />
+    <HashRouter>
+        <App />
+    </HashRouter>
 );
 
 // If you want to start measuring performance in your app, pass a function
